SideBar.vala - Align titles. / StatusBar.vala - lines → sentences.
In SideBar.vala, I've fixed clicking on entries to jump to them, and
made the section titles align with the start of what they're heading
of.

In Statusbar.vala, I've changed the algorithm to count sentences
instead of lines. Also chjanged the algorithm of Reading Time to be
more natural.