SideBar / SideBarBox / HeaderBar / styling - changes all around
In SideBarBox, now the first header of each selected document should
show up and the unselected ones should have a centered title.
It also has a functioning close button again (was a stray if block).
And also, the outline should populate as you type things.

In SideBar, it now controls more intelligently the title and subtitle
of the app.

In HeaderBar, the color buttons in the main menu should have no
strange styling bugs.

Also, the statusbar now is shown by default instead of hidden,
and can still be hidden if the user wants.

Fixes #369, #370, #371, #372