HeaderBar.vala - Move Statusbar components here / Add Save button
The previously located components in the statusbar, were all moved
to the headerbar, with intuitive styling.
Also, there's an explicit Save button, with a dropdown by its side
where Save As and the Export options are now located.
This solves the unintuitiveness of the Saving functions, while
making the UI a bit more cohesive.